+++ derived/src/util.rs
@@ -10,7 +10,9 @@ gen_typeset! {
 
 /// Returns the field names and their corresponding type from the AST (returning an error
 /// if it isn't a struct)
-pub fn get_struct_field_names(ast: &DeriveInput) -> Result<Vec<(Ident, Type)>, TokenStream> {
+pub fn get_struct_field_names(
+    ast: &DeriveInput,
+) -> Result<Vec<(&Ident, &Type, &Vec<Attribute>)>, TokenStream> {
     let fields = match &ast.data {
         Data::Struct(DataStruct {
             fields: Fields::Named(fields),
@@ -30,19 +32,17 @@ pub fn get_struct_field_names(ast: &DeriveInput) -> Result<Vec<(Ident, Type)>, T
         Ok(fields
             .iter()
             .map(|field| {
-                let ty = field.ty.clone();
-                let fname = field.ident.as_ref().unwrap().clone();
-                (fname, ty)
+                let fname = field.ident.as_ref().unwrap();
+                (fname, &field.ty, &field.attrs)
             })
             .collect())
     }
 }
 
-/// Returns a const-ed (if required) func "header"
-pub fn get_func_header(
+pub(crate) fn single_instance_of_attr(
     attrs: &[Attribute],
     target: &str,
-) -> Result<quote::__private::TokenStream, TokenStream> {
+) -> Result<bool, TokenStream> {
     let mut span = None;
     let has_attr = attrs
         .iter()
@@ -52,12 +52,8 @@ pub fn get_func_header(
         })
         .count();
     match has_attr {
-        0 => Ok(quote! {
-            pub fn
-        }),
-        1 => Ok(quote! {
-            pub const fn
-        }),
+        0 => Ok(false),
+        1 => Ok(true),
         _ => Err(syn::Error::new(
             span.unwrap(),
             format!("Found duplicate attributes for `{}`", target),
@@ -65,4 +61,21 @@ pub fn get_func_header(
         .into_compile_error()
         .into()),
     }
+}
+
+/// Returns a const-ed (if required) func "header"
+pub(crate) fn get_func_header(
+    attrs: &[Attribute],
+    target: &str,
+) -> Result<quote::__private::TokenStream, TokenStream> {
+    let r = if self::single_instance_of_attr(attrs, target)? {
+        quote! {
+            pub const fn
+        }
+    } else {
+        quote! {
+            pub fn
+        }
+    };
+    Ok(r)
 }