+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All changes to this project will be noted in this file
 
+## Unreleased
+
+### Additions
+
+- Added the `#[const_gtor]` and `#[const_ctor]` attributes for providing compile-time ctors and getters
+- Added the `#[gtor_copy]` attribute for compound `Copy` types (UDFs)
+
 ## 0.3.1
 
 - Fixed incorrect compiler error messages and inconsistencies in the documentation